backend/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
pillow
apscheduler
sqlalchemy
psycopg2-binary
python-dotenv
python-dotenv>=1.0.0
python-jose
passlib==1.7.4
bcrypt==3.2.0
redis
celery
torch
pydantic[email]
starlette
requests
python-jose[cryptography]==3.3.0